# bufferpool: Improve performance