--- a/exp/bufferpool/pool_test.go
+++ b/exp/bufferpool/pool_test.go
@@ -10,13 +10,8 @@ import (
 func TestInvariants(t *testing.T) {
 	t.Parallel()
 
-	pool := bufferpool.Pool{
-		BucketCount: 1,
-		MinAlloc:    1024,
-	}
-
 	assert.Panics(t, func() {
-		pool.Get(32)
+		bufferpool.NewPool(1024, 1)
 	}, "should panic when BucketCount cannot satisfy MinAlloc.")
 }
 
@@ -24,38 +19,40 @@ func TestGet(t *testing.T) {
 	t.Parallel()
 	t.Helper()
 
-	pool := bufferpool.Pool{
-		BucketCount: 8,
-		MinAlloc:    64,
-	}
+	minAlloc, bucketCount := 64, 8
+	pool := bufferpool.NewPool(minAlloc, bucketCount)
 
 	t.Run("Size<MinAlloc", func(t *testing.T) {
-		assert.Len(t, pool.Get(32), 32, "should return buffer with len=32")
-		assert.Equal(t, 32, cap(pool.Get(32)), "should return buffer with cap=MinAlloc")
+		size := minAlloc / 2 // size < minAlloc
+		assert.Len(t, pool.Get(size), size, "should return buffer with len=%d", size)
+		assert.Equal(t, minAlloc, cap(pool.Get(size)), "should return buffer with cap=MinAlloc")
 	})
 
 	t.Run("Size=MinAlloc", func(t *testing.T) {
-		assert.Len(t, pool.Get(pool.MinAlloc), pool.MinAlloc, "should return buffer with len=MinAlloc")
-		assert.Equal(t, pool.MinAlloc, cap(pool.Get(pool.MinAlloc)), "should return buffer with cap=MinAlloc")
+		assert.Len(t, pool.Get(minAlloc), minAlloc, "should return buffer with len=MinAlloc")
+		assert.Equal(t, minAlloc, cap(pool.Get(minAlloc)), "should return buffer with cap=MinAlloc")
 	})
 
 	t.Run("Size>MinAlloc", func(t *testing.T) {
-		assert.Len(t, pool.Get(33), 33, "should return buffer with len=33")
-		assert.Equal(t, 128, cap(pool.Get(128)), "should return buffer with cap=128")
+		size := minAlloc + 1
+		nextAlloc := minAlloc << 1
+		assert.Len(t, pool.Get(size), size, "should return buffer with len=%d", size)
+		assert.Equal(t, nextAlloc, cap(pool.Get(size)), "should return buffer with cap=%d", nextAlloc)
 	})
 
 	t.Run("Size>MaxSize", func(t *testing.T) {
-		assert.Len(t, pool.Get(512), 512, "should return buffer with len=512")
-		assert.GreaterOrEqual(t, 512, cap(pool.Get(512)), "should return buffer with cap>=512")
+		size := 1<<(bucketCount+1) + 1
+		assert.Len(t, pool.Get(size), size, "should return buffer with len=%d", size)
+		assert.GreaterOrEqual(t, size, cap(pool.Get(size)), "should return buffer with cap>=%d", size)
 	})
 }
 
 func TestPut(t *testing.T) {
 	t.Parallel()
 
-	var pool bufferpool.Pool
+	pool := bufferpool.NewPool(1024, 20)
 
-	buf := make([]byte, 1024)
+	buf := make([]byte, 1024, 1024)
 	for i := range buf {
 		buf[i] = byte(i)
 	}
@@ -63,12 +60,15 @@ func TestPut(t *testing.T) {
 
 	pool.Put(buf)
 	buf = pool.Get(8)
+	fullBuf := buf[:cap(buf)]
 
+	assert.Len(t, fullBuf, 1024, "buf should have len and cap 1024")
 	assert.Equal(t, make([]byte, 8), buf, "should zero first 8 bytes")
+	assert.NotEqual(t, 0, fullBuf[9], "first byte after clearing should not be zero")
 }
 
 func BenchmarkPool(b *testing.B) {
-	var pool bufferpool.Pool
+	pool := bufferpool.NewPool(0, 0)
 	const size = 32
 
 	// Make cache hot.