@matheusd 2025-01-11
I haven't looked, but two things to watch for:

1. I believe there was actually a bug in the prior bufferpool implementation that was not respecting `minAlloc`, so using this fixed version increased allocs (and thus reduced overall throughput).

2. The benchmark may not be releasing data back to the bufferpool, which combined with the prior point may be increasing pressure to runtime memory management (that was the reason I added the `.Release()` in the prior benchmark).